Make third example register 16-bit for Python

+++ test/sampleData/arduino/Example.h
@@ -71,16 +71,16 @@ class Example {
 
          */
         int writeRegisterB(uint16_t data);        /**
-         * A 32-bit register
+         * A 16-bit register
 
          */
-        uint32_t readRegisterC();
+        uint16_t readRegisterC();
 
         /**
-         * A 32-bit register
+         * A 16-bit register
 
          */
-        int writeRegisterC(uint32_t data);        /**
+        int writeRegisterC(uint16_t data);        /**
          * This is a few bits
 
          */
